SignOpCache
This PR introduces a SignOpCache, which gathers intermediate data
obtained during signature validation (signatures, public keys, sig-
hashes, subscript) so it can be inspected later.

SignOpCache was written to tolerate multiple CHECKSIG/CHECKMULTISIG
opcodes throughout a script. To create a SignOpCache for a script,
call Engine.ExecuteSignOp(allowIncomplete), - AllowIncomplete is
used to obtain the state of an incomplete signature operation if
it is expected.

The motivation for this PR was initially the requirement of assoc-
iating keys and sigs in a partially/completely signed transaction.